day_39_flight_deal_finder/requirements.txt
requests~=2.32.3
python-dotenv~=1.0.1
twilio~=9.4.4